[chip_sw, kmac_entropy] Checking timeout of EDN to KMAC
Note: this is the first incremental towards kmac_entropy test, however
there is an issue with the verilator.

This commit includes a simple test where ENTROPY_PERIOD reg is
configured such that:

`prescaler` = 0x3FF
`wait_timer` = 0xFFFF

so that we can observe the maximum timeout value for the EDN request.
The issue is that even though we pick the largest possible values, EDN
still times out (which should not happen).

In the case of FPGA, the EDN request does not time out.

Signed-off-by: Fatih Balli <fatihballi@opentitan.org>